Make SharedTransitionLayer work. Total Refactor of App Navigation
