--- app/src/main/java/com/rabbitv/valheimviki/navigation/ChildNavGraph.kt
@@ -1,28 +1,0 @@
-package com.rabbitv.valheimviki.navigation
-
-import androidx.compose.animation.AnimatedVisibilityScope
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.padding
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.composable
-import com.rabbitv.valheimviki.presentation.biome.BiomeScreen
-import com.rabbitv.valheimviki.presentation.creatures.bosses.BossScreen
-import com.rabbitv.valheimviki.presentation.creatures.mini_bosses.MiniBossScreen
-
-
-fun NavGraphBuilder.addHomeGraph(
-    paddingValues: PaddingValues,
-    valheimVikiNavController: NavHostController,
-    animatedVisibilityScope: AnimatedVisibilityScope
-) {
-
-}
-
-
-
-
-
-
